Crate popcap_pak

Structs

Entry

An Entry, representative of a file inside a pakfile.

FileTime

A file time.

Pak

An in-memory PAK file.

Enums

PakError

Error type of this library

TryFromFileTimeError

A failure occured while converting from a file time into a system time

TryFromSystemTimeError

An error that may occur while converting a SystemTime into a FileTime.